Exam Question: add api details exam question bank

+++ lms-v2/lmsbackend/baseweb/src/main/java/com/hust/baseweb/applications/exam/service/impl/ExamQuestionServiceImpl.java
@@ -2,15 +2,19 @@ package com.hust.baseweb.applications.exam.service.impl;
 
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.hust.baseweb.applications.contentmanager.repo.MongoContentService;
 import com.hust.baseweb.applications.exam.entity.ExamEntity;
 import com.hust.baseweb.applications.exam.entity.ExamQuestionEntity;
+import com.hust.baseweb.applications.exam.entity.ExamTestQuestionEntity;
 import com.hust.baseweb.applications.exam.model.ResponseData;
 import com.hust.baseweb.applications.exam.model.request.ExamQuestionDeleteReq;
 import com.hust.baseweb.applications.exam.model.request.ExamQuestionDetailsReq;
 import com.hust.baseweb.applications.exam.model.request.ExamQuestionFilterReq;
 import com.hust.baseweb.applications.exam.model.request.ExamQuestionSaveReq;
 import com.hust.baseweb.applications.exam.repository.ExamQuestionRepository;
+import com.hust.baseweb.applications.exam.repository.ExamTestQuestionRepository;
 import com.hust.baseweb.applications.exam.service.ExamQuestionService;
+import com.hust.baseweb.applications.exam.service.MongoFileService;
 import com.hust.baseweb.applications.exam.utils.DataUtils;
 import com.hust.baseweb.applications.exam.utils.SecurityUtils;
 import lombok.RequiredArgsConstructor;
@@ -21,6 +25,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
@@ -35,6 +40,8 @@ import java.util.Optional;
 public class ExamQuestionServiceImpl implements ExamQuestionService {
 
     private final ExamQuestionRepository examQuestionRepository;
+    private final ExamTestQuestionRepository examTestQuestionRepository;
+    private final MongoFileService mongoFileService;
     private final ModelMapper modelMapper;
     private final EntityManager entityManager;
     private final ObjectMapper objectMapper;
@@ -112,7 +119,7 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
     }
 
     @Override
-    public ResponseData<ExamQuestionEntity> create(ExamQuestionSaveReq examQuestionSaveReq) {
+    public ResponseData<ExamQuestionEntity> create(ExamQuestionSaveReq examQuestionSaveReq, MultipartFile[] files) {
         ResponseData<ExamQuestionEntity> responseData = new ResponseData<>();
         Optional<ExamQuestionEntity> examQuestionExist = examQuestionRepository.findByCode(examQuestionSaveReq.getCode());
         if(examQuestionExist.isPresent()){
@@ -122,7 +129,10 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             return responseData;
         }
 
+        List<String> filePaths = mongoFileService.storeFiles(files);
+
         ExamQuestionEntity examQuestionEntity = modelMapper.map(examQuestionSaveReq, ExamQuestionEntity.class);
+        examQuestionEntity.setFilePath(String.join(";", filePaths));
         examQuestionRepository.save(examQuestionEntity);
         responseData.setHttpStatus(HttpStatus.OK);
         responseData.setResultCode(HttpStatus.OK.value());
@@ -132,7 +142,7 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
 
     @Override
     @Transactional
-    public ResponseData<ExamQuestionEntity> update(ExamQuestionSaveReq examQuestionSaveReq) {
+    public ResponseData<ExamQuestionEntity> update(ExamQuestionSaveReq examQuestionSaveReq, MultipartFile[] files) {
         ResponseData<ExamQuestionEntity> responseData = new ResponseData<>();
         Optional<ExamQuestionEntity> examQuestionExist = examQuestionRepository.findByCode(examQuestionSaveReq.getCode());
         if(!examQuestionExist.isPresent()){
@@ -142,11 +152,19 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             return responseData;
         }
 
+        List<String> filePaths = mongoFileService.storeFiles(files);
+
         ExamQuestionEntity examQuestionEntity = modelMapper.map(examQuestionSaveReq, ExamQuestionEntity.class);
         examQuestionEntity.setId(examQuestionExist.get().getId());
         examQuestionEntity.setCreatedBy(examQuestionExist.get().getCreatedBy());
         examQuestionEntity.setCreatedAt(examQuestionExist.get().getCreatedAt());
+        examQuestionEntity.setFilePath(examQuestionEntity.getFilePath() +";"+String.join(";", filePaths));
         examQuestionRepository.save(examQuestionEntity);
+
+        for(String filePath: examQuestionSaveReq.getDeletePaths()){
+            mongoFileService.deleteByPath(filePath);
+        }
+
         responseData.setHttpStatus(HttpStatus.OK);
         responseData.setResultCode(HttpStatus.OK.value());
         responseData.setResultMsg("Cập nhật câu hỏi thành công");
@@ -164,6 +182,20 @@ public class ExamQuestionServiceImpl implements ExamQuestionService {
             responseData.setResultMsg("Chưa tồn tại câu hỏi");
             return responseData;
         }
+
+        List<ExamTestQuestionEntity> examTestQuestionEntityList = examTestQuestionRepository.findAllByExamQuestionId(examQuestionDeleteReq.getId());
+        if(!examTestQuestionEntityList.isEmpty()){
+            responseData.setHttpStatus(HttpStatus.NOT_FOUND);
+            responseData.setResultCode(HttpStatus.NOT_FOUND.value());
+            responseData.setResultMsg("Câu hỏi đã được gán cho đề thi, không được xoá");
+            return responseData;
+        }
+
+        String[] filePaths = examQuestionExist.get().getFilePath().split(";");
+        for(String filePath: filePaths){
+            mongoFileService.deleteByPath(filePath);
+        }
+
         examQuestionRepository.delete(examQuestionExist.get());
         responseData.setHttpStatus(HttpStatus.OK);
         responseData.setResultCode(HttpStatus.OK.value());
